--- FE/src/components/Issues/IssueTable/IssueList/AvatarStack/Avatar.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Avatar = () => {
-  return <div />;
-};
-
-export default Avatar;
