--- FE/src/components/Issues/IssueTable/IssueList/AvatarStack/AvatarStack.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const AvatarStack = () => {
-  return <div />;
-};
-
-export default AvatarStack;
