+++ FE/src/components/Issues/IssueTable/IssueList/Contents/Details.jsx
@@ -1,7 +1,14 @@
 import React from "react";
 
+import Box from "@material-ui/core/Box";
+import Typography from "@material-ui/core/Typography";
+
 const Details = () => {
-  return <div />;
+  return (
+    <Box color="text.secondary">
+      <Typography variant="body2">#18 opened 4 days ago by sungik-choi</Typography>
+    </Box>
+  );
 };
 
 export default Details;
