--- FE/src/components/Issues/IssueTable/IssueList/Contents/Label.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Label = () => {
-  return <div />;
-};
-
-export default Label;
