+++ FE/src/components/Issues/IssueTable/IssueList/Contents/Labels.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+import Chip from "@material-ui/core/Chip";
+
+// ! label : 레이블 내용 ! color : 레이블 컬러 (별도의 스타일 지정 필요)
+
+const Labels = () => {
+  return <Chip label="FE" size="small" color="primary" />;
+};
+
+export default Labels;
