+++ FE/src/components/Issues/IssueTable/IssueList/StatusIcon.jsx
@@ -1,7 +1,11 @@
 import React from "react";
 
+import ErrorOutlineIcon from "@material-ui/icons/ErrorOutline";
+
+// ! 이슈 Close 여부에 따라 아이콘 변경 필요
+
 const StatusIcon = () => {
-  return <div />;
+  return <ErrorOutlineIcon color="primary" />;
 };
 
 export default StatusIcon;
